+++ test/test_helper.rb
@@ -9,3 +9,17 @@ begin
 rescue LoadError
   HAVE_RBNACL = false
 end
+
+begin
+  require "digest/blake3"
+  HAVE_BLAKE3 = true
+rescue LoadError
+  HAVE_BLAKE3 = false
+end
+
+begin
+  require "chacha20blake3"
+  HAVE_CHACHA20BLAKE3 = true
+rescue LoadError
+  HAVE_CHACHA20BLAKE3 = false
+end